requirements.txt
langchain>=0.3.0
langchain-anthropic>=0.3.0
langchain-community>=0.3.0
anthropic>=0.40.0
rich>=13.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
